+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### unreleased
 - **add:** expose localhost
+- **add:** test files for header and validations
 
 ### [v0.2.1] - 2016-06-26
 - **fix:** refactor file structure