+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### unreleased
 - **add:** expose localhost
 - **add:** test files for header and validations
+- **fix:** doc block
 
 ### [v0.2.1] - 2016-06-26
 - **fix:** refactor file structure