+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **add:** expose localhost
 - **add:** test files for header and validations
 - **fix:** doc block
+- **rmv:** left console logs
 
 ### [v0.2.1] - 2016-06-26
 - **fix:** refactor file structure