+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **add:** test files for header and validations
 - **fix:** doc block
 - **rmv:** left console logs
+- **doc:** fix flaw
 
 ### [v0.2.1] - 2016-06-26
 - **fix:** refactor file structure