+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - **fix:** doc block
 - **rmv:** left console logs
 - **doc:** fix flaw
+- **fix:** exposing of result
 
 ### [v0.2.1] - 2016-06-26
 - **fix:** refactor file structure