+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - **fix:** exposing of result
 - **add:** tests for validate.options
 - **add:** tests for validate.query
+- **fix:** tests for validate.query
 
 ### [v0.2.1] - 2016-06-26
 - **fix:** refactor file structure