+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - **add:** tests for validate.query
 - **fix:** tests for validate.query
 - **rmv:** left console logs
+- **fix:** unset link header if no links are available
 
 ### [v0.2.1] - 2016-06-26
 - **fix:** refactor file structure